更新抢购逻辑，更新sku_detail表

+++ src/main/java/team/naive/secondkillsaas/Biz/OrderService.java
@@ -16,7 +16,7 @@ public interface OrderService {
 
     OrderDetailBO getOrderDetail(Long orderId);
 
-    OrderBO getOrderDetail(Long userId, Long skuId);
+    OrderBO getOrderDetail(Integer userId, Long skuId);
 
     ResponseVO cancelOrder(Long orderId);
